Little more abstraction started, fixed a couple more errors, progress is slow but speeding up for sure.
In that abstraction, I'm considering being able to register classes to fix, Runnables(Functions, Predicates, BiFunctions, that kind of thing) to transform or mark for removal single or a number of instructions.

Either way, in InstructionFixers I added InstructionFixers#applyStaticInsnClassMoves and InstructionFixers#ONE_TO_ONE_CLASS_MOVES to do one-to-one class movements, for instance from the wrong VecHelper to the right one.

Also added ClassConstants for enumerating constants useful for all of this.