Did a bunch! like. a lot.
Got it almost to the point of running, I think this includes all of the Visuals and ActorVisuals as well as a lot of other renderer fixes. Working on Ponder stuff right now, for some reason it's kinda broken with DDBlocks#ponder_stone_generation but I'll fix that later.

Next is to finish fixing VoxelShapes!

Got a couple of things still to fix and to test, mainly the visuals' registration which is kinda hacky, it's in AddInstanceToCreateBlockEntityBuilder, but I haven't gotten to the point where those'd be testable...

Still need to do placement helpers as well, but I'm getting closer and closer!